reformat sql statements in constructor service
