+++ gladanalysis/services/__init__.py
@@ -6,6 +6,6 @@ from __future__ import print_function
 
 from gladanalysis.services.geostore_service import GeostoreService
 from gladanalysis.services.date_service import DateService
-from gladanalysis.services.sql_service import SqlService
+from gladanalysis.services.query_constructor_service import QueryConstructorService
 from gladanalysis.services.analysis_service import AnalysisService
 from gladanalysis.services.response_service import ResponseService
